+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.15.0
+## Add
+  - `Name` field on FlowStart model.
+
+## Change
+  - Restrict room flow start creation to only one active flow start per room(only when starting with a room).
+
+
+
 # 1.0.1
 ## Add
   - Changelog init.